requirements.txt
clickclick>=1.0
requests
stups-zign>=1.0.10
stups-cli-support
